fix: tag.phpのタグページ(.../cat/...)を修正

+++ wordpress/wp-content/themes/vivid/tag.php
@@ -49,7 +49,7 @@
                                             <span class="c-lisence-card__section acji-item-name"><?php the_title(); ?></span>
                                         </h3>
                                         <p class="c-lisence-card__content">
-                                            <s class="c-lisence-card__section"><?php echo number_format(get_field('list_price')); ?>円</s>
+                                            <s class="c-lisence-card__section"><?php echo number_format((float)get_field('list_price')); ?>円</s>
                                             <span class="acji-item-off u-font-color--tomato"></span>
                                             <span class="c-lisence-card__price acji-item-price"></span>
                                         </p>
